CH-1/pyproject.toml
[project]
name = "ch-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-google-genai>=4.2.0",
]
